pyproject.toml
[tool.poetry]
name = "codebook"
version = "0.1.0"
description = ""
authors = ["MrGigSolutions <mark@gig.solutions>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
